Добавит компонент FavoriteEmpty
