+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,81 +1,19 @@
 import { Link } from 'react-router-dom';
-import PlaceCard from '../../components/place-card/place-card';
-import { offers } from '../../mocks/offers';
+import FavoriteContainer from '../../components/favorite-container/favorite-container';
+import FavoriteEmpty from '../../components/favorite-empty/favorite-empty';
+import { Offer } from '../../types/types';
 
+type FavoritesPageProps = {
+  favorites: Offer[];
+}
 
-function FavoritesPage(): JSX.Element {
-  const favorites = offers.filter((item) => item.isFavorite);
+function FavoritesPage({favorites}: FavoritesPageProps): JSX.Element {
 
   return (
-    <div className="page">
-
-      <main className="page__main page__main--favorites">
+    <div className={`page ${favorites.length < 1 ? 'page--favorites-empty' : ''}`}>
+      <main className={`page__main page__main--favorites ${favorites.length < 1 ? 'page__main--favorites-empty' : ''}`}>
         <div className="page__favorites-container container">
-          <section className="favorites">
-            <h1 className="favorites__title">Saved listing</h1>
-            <ul className="favorites__list">
-              <li className="favorites__locations-items">
-                <div className="favorites__locations locations locations--current">
-                  <div className="locations__item">
-                    <a className="locations__item-link" href="#">
-                      <span>Amsterdam нужен фильтр по городам</span>
-                    </a>
-                  </div>
-                </div>
-                <div className="favorites__places">
-                  <article className="favorites__card place-card">
-                    <div className="place-card__mark">
-                      <span>Premium</span>
-                    </div>
-                    <div className="favorites__image-wrapper place-card__image-wrapper">
-                      <a href="#">
-                        <img className="place-card__image" src="img/apartment-small-03.jpg" width="150" height="110" alt="Place image" />
-                      </a>
-                    </div>
-                    <div className="favorites__card-info place-card__info">
-                      <div className="place-card__price-wrapper">
-                        <div className="place-card__price">
-                          <b className="place-card__price-value">&euro;180</b>
-                          <span className="place-card__price-text">&#47;&nbsp;night</span>
-                        </div>
-                        <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
-                          <svg className="place-card__bookmark-icon" width="18" height="19">
-                            <use xlinkHref="#icon-bookmark"></use>
-                          </svg>
-                          <span className="visually-hidden">In bookmarks</span>
-                        </button>
-                      </div>
-                      <div className="place-card__rating rating">
-                        <div className="place-card__stars rating__stars">
-                          <span style={{width: '100%'}}></span>
-                          <span className="visually-hidden">Rating</span>
-                        </div>
-                      </div>
-                      <h2 className="place-card__name">
-                        <a href="#">Nice, cozy, warm big bed apartment</a>
-                      </h2>
-                      <p className="place-card__type">Apartment</p>
-                    </div>
-                  </article>
-                </div>
-              </li>
-
-              <li className="favorites__locations-items">
-                <div className="favorites__locations locations locations--current">
-                  <div className="locations__item">
-                    <a className="locations__item-link" href="#">
-                      <span>Cologne</span>
-                    </a>
-                  </div>
-                </div>
-                <div className="favorites__places">
-
-                  {favorites.map((item) => <PlaceCard key={item.id} offer={item} className='favorites'/>)}
-
-                </div>
-              </li>
-            </ul>
-          </section>
+          {favorites.length < 1 ? <FavoriteEmpty /> : <FavoriteContainer favorites={favorites}/>}
         </div>
       </main>
       <footer className="footer container">
